refactored, so it's easier to follow
